replace eufy-node-client wiht eufy-security-client
Main driver behind this change are issues with Firebase registration - see
https://github.com/matijse/eufy-ha-mqtt-bridge/issues/78 for details.

Aside from different name, the main change is how the registration is handled.
Eufy-security-client's new `PushNotificationService` class handles the
registration itself if needed and emits the 'credential' event when it happens.
On our side, we just listen to this event and persist credentials to disk.

In future, we may want to stop listening to raw events and use normalized
messages provided by `PushNotificationService` instead. This would potentially
allow supporting more devices.